Move + better doc swapoff --all in legacy rc.local
Moves the use of "swapoff --all" in the "rc.local" file to the top-of-the-file
as a very quick defensive workaround to the '-e' exit-on-first-error option.

Documents in the "rc.local" file the intention to deprecate  "rc.local"
entirely left for a bigger refactor in #575.

Additionally updates the FIXME around deprecating the call to 'swapoff --all'
to note my recent attempts to disable 'systemd-gpt-auto-generator' (which
includes the swap mounting code) with the kernel boot
option'systemd.gpt_auto=no', and masking the 'swap.target' unit. Both of which
did not succeed in deprecating 'swapoff --all'.

Issue: #515, #565
Signed-off-by: engstk <eng.stk@sapo.pt>